new model/requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
numpy
joblib
python-dotenv
google-generativeai
